Add Detectron2ModelCatalog with COCO-Detection models.
