+++ MANIFEST.in
@@ -1,0 +1,2 @@
+include configs/Base-*.yaml
+include configs/COCO-Detection/*.yaml
