+++ detectron2/checkpoint/__init__.py
@@ -3,7 +3,7 @@
 # File:
 
 
-from . import model_zoo as _UNUSED  # register the handler
+from . import catalog as _UNUSED  # register the handler
 from .detection_checkpoint import DetectionCheckpointer
 from fvcore.common.checkpoint import Checkpointer, PeriodicCheckpointer
